Unfinished of course. Adding boilerplate app code so I can scaffold an API ready crud app in minutes
What would it be cool to do?
- Be able to define an API from a yaml file and bring it in?
- Be able to define domain objects

add an --ecommerce flag to add in basket component, checkout component

add an --account flag to add in the login, logout and all other auth
stuff with an account homepage

routes.

Also, do I want to throw in an express application and mongodb
connection?! that'd be amazing.

AND testing